Adds publication's information to the profile
Now, the profile has two lists, one for publications from conferences and one for publications from magazines. Both are dicts with keys like: 'Tipo', 'Título da Publicação', 'Fórum', 'Ano' and 'Páginas'. If the publication is from a magazine, it's possible for it to have a 'ISSN' as key too.