The entities of a invalid report are written as new worksheets
If a report has more than one entity, other than researcher, and reports_as_new_worksheets is set to False, the user can set new_worksheet_if_conflict to True and the entities in a report are going to be written in a new worksheet each. But the data about researchers are going to be written in all worksheets.